requirements.txt
django ==2.2.26
psycopg2-binary == 2.9.5
Faker == 4.1.1
djangorestframework == 3.9.0
django-cors-headers == 3.2.0
gunicorn== 20.1.0
django-utils-six ==2.0
whitenoise ==3.3.1
python-decouple ==3.8
dj-database-url==0.5.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
Pillow==9.4.0